requirement.txt
astroid==2.1.0
bootstrap-admin==0.4.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
Django==2.1.5
django-cors-headers==2.4.0
django-filter==2.0.0
django-markdownx==2.0.25
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
docutils==0.14
drf-yasg==1.12.1
idna==2.8
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
Pillow==5.3.0
PyJWT==1.7.1
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
pytz==2018.7
requests==2.21.0
ruamel.yaml==0.15.84
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.11.0
